Provide only one text object (i)
Provide only one text object behaving differently depending on the
current filetype. It doesn't include the line below the indent text for
file types where indentantion is a part of the syntax. Currently it
means Python, CoffeeScript, and Haml buffers. This can be extended as
necessary.

If accepted the doc/indent-object.txt should be changed accordingly.