Fix issue arising when used with remapped keys.
This fixes an issue arising from use of normal instead of normal! inside
the text object function. normal! is used in all places except one, the
final gv. The result is that if a user has remapped built in VIM
commands (gv, in this case), indent objects will not work correctly.