Use xnoremap instead of vnoremap
The `:vnoremap` command affects both Visual and Select modes, but the
Visual mode bindings provided break in Select mode, since the `gv` is
interpreted as replacing the selection with the literal characters "gv".
Since this binding is probably not desirable in Select mode in any case,
I propose simply changing the commands to use `:xnoremap`, which applies
only to Visual mode (see `:help mapmode-x`).